+++ cmsdev/internal/test/bos/bos_sessions_cli.go
@@ -31,6 +31,7 @@ package bos
 import (
 	"encoding/json"
 	"strconv"
+	"strings"
 
 	"stash.us.cray.com/SCMS/cms-tools/cmsdev/internal/lib/common"
 )
@@ -46,6 +47,11 @@ func CreateBOSSessionCLI(staged bool, sessionName, templateName, cliVersion stri
 			common.Error(err)
 		}
 	}
+	// If the tenant is a dummy tenant, we expect the command to fail
+	if strings.Contains(common.GetTenantName(), "dummy-tenant") {
+		return BOSSession{}, true
+	}
+
 	return
 }