+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- CASMCMS-9472: Add multitenancy BOS CRUD tests to cmsdev
+
 ## [1.33.0] - 2025-06-27
 
 ### Changed